docs(styleguidist-overrides): Separate styleguidist overrides from custom components
* Split docs-only components into custom/ and overrides/
* Turn off prettier and linting for the overrides, as they follow the
styleguidist conventions
