+++ .eslintignore
@@ -1,2 +1,3 @@
 scripts/scaffolding
 dist
+docs/components/overrides
