+++ .prettierignore
@@ -1,1 +1,2 @@
 package.json
+docs/components/overrides
