+++ docs/components/custom/Markdown/Markdown.js
@@ -5,20 +5,20 @@ import PropTypes from 'prop-types'
 import { compiler } from 'markdown-to-jsx'
 import mapValues from 'lodash/mapValues'
 // import memoize from 'lodash/memoize';
-import Styled from 'react-styleguidist/lib/rsg-components/Styled'
+import Styled from 'rsg-components/Styled'
 
-import { styles as paraStyles } from 'react-styleguidist/lib/rsg-components/Para'
+import { styles as paraStyles } from 'rsg-components/Para'
 
-import Link from '../../../packages/Link/Link'
-import Text from '../../../packages/Text/Text'
-import Strong from '../../../packages/Strong/Strong'
+import Link from '../../../../packages/Link/Link'
+import Text from '../../../../packages/Text/Text'
+import Strong from '../../../../packages/Strong/Strong'
 import MarkdownHeading from '../MarkdownHeading/MarkdownHeading'
 import MarkdownParagraph from '../MarkdownParagraph/MarkdownParagraph'
 import MarkdownUnorderedList from '../MarkdownUnorderedList/MarkdownUnorderedList'
 import MarkdownOrderedList from '../MarkdownOrderedList/MarkdownOrderedList'
 
 // loading our own custom version of highlight.js docco theme to make code examples AAA colour accessible
-import codeStyles from '../../css/highlight_js/docco.css'
+import codeStyles from './highlight_js/docco.css'
 
 // Temporary disable memoization to fix: https://github.com/styleguidist/react-styleguidist/issues/348
 // TODO: Remove after merge: https://github.com/probablyup/markdown-to-jsx/pull/96
