+++ docs/components/custom/MarkdownHeading/MarkdownHeading.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Box from '../../../packages/Box/Box'
-import Heading from '../../../packages/Heading/Heading'
+import Box from '../../../../packages/Box/Box'
+import Heading from '../../../../packages/Heading/Heading'
 
 const MarkdownHeading = ({ level, children }) => (
   <Box below={3}>
