+++ docs/components/custom/MarkdownOrderedList/MarkdownOrderedList.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Box from '../../../packages/Box/Box'
-import OrderedList from '../../../packages/OrderedList/OrderedList'
-import Text from '../../../packages/Text/Text'
+import Box from '../../../../packages/Box/Box'
+import OrderedList from '../../../../packages/OrderedList/OrderedList'
+import Text from '../../../../packages/Text/Text'
 
 const MarkdownOrderedList = ({ children }) => (
   <Box below={3}>
