+++ docs/components/custom/MarkdownParagraph/MarkdownParagraph.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Box from '../../../packages/Box/Box'
-import Paragraph from '../../../packages/Paragraph/Paragraph'
+import Box from '../../../../packages/Box/Box'
+import Paragraph from '../../../../packages/Paragraph/Paragraph'
 
 const MarkdownParagraph = ({ small, children }) => (
   <Box below={3}>
