+++ docs/components/custom/MarkdownUnorderedList/MarkdownUnorderedList.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Box from '../../../packages/Box/Box'
-import UnorderedList from '../../../packages/UnorderedList/UnorderedList'
-import Text from '../../../packages/Text/Text'
+import Box from '../../../../packages/Box/Box'
+import UnorderedList from '../../../../packages/UnorderedList/UnorderedList'
+import Text from '../../../../packages/Text/Text'
 
 const MarkdownUnorderedList = ({ children }) => (
   <Box below={3}>
