+++ docs/components/custom/SectionHeading/SectionHeadingRenderer.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Heading from '../../../packages/Heading/Heading'
-import Flexbox from '../../../shared/components/Flexbox/Flexbox'
-import Box from '../../../packages/Box/Box'
+import Heading from '../../../../packages/Heading/Heading'
+import Flexbox from '../../../../shared/components/Flexbox/Flexbox'
+import Box from '../../../../packages/Box/Box'
 
 import styles from './SectionHeadingRenderer.modules.scss'
