+++ docs/components/custom/SectionHeading/SectionHeadingRenderer.modules.scss
