+++ docs/components/custom/SectionHeading/__tests__/SectionHeading.spec.jsx
