+++ docs/components/overrides/Editor/codemirror/ttcn.css
