+++ package.json
@@ -131,7 +131,7 @@
   },
   "lint-staged": {
     "*.{js,jsx}": [
-      "eslint --ext '.js,.jsx' --config config/eslint.config.js --max-warnings 0 --fix",
+      "eslint --ext '.js,.jsx' --config config/eslint.config.js --quiet --fix",
       "git add"
     ],
     "*.scss": "stylelint --config config/stylelint.config.js",
